requirements.txt
scikit-learn==1.0.2
tensorflow==2.17.0
numpy
pillow==10.3.0